Safari-611.1.10/LayoutTests/http/tests/plugins/supported-plugin-all-origin-visibility.html

<html>
<body>
<script src= "../resources/testharness.js"></script>
<script src= "../resources/testharnessreport.js"></script>
<script>
function checkPlugins()
{
    return !!(navigator.plugins["WebKit Test PlugIn"] || navigator.plugins["WebKit built-in PDF"]);
}

test(() => {
    assert_true(!!window.testRunner);
    testRunner.setPluginSupportedMode("allOrigins");
    var plugins = navigator.plugins;
    assert_true(checkPlugins());
}, "Ensure plugins visibility for supported plug-in for all origins");
</script>
</body>
</html>
